Fix: "권한 변경 클릭 수정"
권한 변경 클릭 시 페이지 리로딩해서 방 정보 바로 확인할 수 있도록 수정

+++ src/components/userdetail/UserDetailMain.vue
@@ -114,15 +114,42 @@ export default {
         withCredentials: true
       })
           .then(response => {
-            window.alert("권한 변경 성공");
-            this.userInfo = JSON.parse(localStorage.getItem("userInfo"));
-            this.userInfo.role = "HOST";
-            this.isUser = false;
-            localStorage.setItem("userInfo",JSON.stringify(this.userInfo));
-            console.log(response.data);
+
+            this.$swal.fire({
+              title: "호스트로 변경하시겠습니까?",
+              text: "호스트 권한은 되돌릴 수 없습니다!",
+              icon: "warning",
+              showCancelButton: true,
+              confirmButtonColor: "#3085d6",
+              cancelButtonColor: "#d33",
+              confirmButtonText: "네",
+              cancelButtonText: "아니오"
+            }).then((swalResult) => {
+              if (swalResult.isConfirmed) {
+
+                this.userInfo = JSON.parse(localStorage.getItem("userInfo"));
+                this.userInfo.role = "HOST";
+                this.isUser = false;
+                localStorage.setItem("userInfo",JSON.stringify(this.userInfo));
+                console.log(response.data);
+
+                this.$swal.fire({
+                    title: "권한 변경에 성공했습니다!",
+                    icon: "success"
+                }).then(() => {
+                  router.go()
+                })
+                ;
+
+              }
+            });
+
           })
           .catch(error => {
-            window.alert("권한 변경 실패");
+            this.$swal.fire({
+                title: "권한 변경에 실패했습니다.",
+                icon: "error"
+            });
             console.error(error);
           });
     },
